+++ ecosystem.config.cjs
@@ -3,17 +3,18 @@ module.exports = {
 		{
 			name: "gamgee",
 			script: "./dist/server.js",
+			node_args: ["--env-file=.env"],
 			cwd: __dirname,
 			source_map_support: true,
 			watch: ["dist"],
 			watch_delay: 1000,
 			time: true,
 			env: {
-				NODE_ENV: "development"
+				NODE_ENV: "development",
 			},
 			env_production: {
-				NODE_ENV: "production"
-			}
-		}
-	]
+				NODE_ENV: "production",
+			},
+		},
+	],
 };