+++ src/actions/network/getSoundCloudTrack.ts
@@ -1,23 +1,28 @@
-import type { VideoDetails } from "../getVideoDetails.js";
+import type { VideoDetails, VideoMetaSource } from "../getVideoDetails.js";
 import type { Song } from "soundcloud-scraper";
 import { Client as SoundCloudClient } from "soundcloud-scraper";
-import { fetchWithTimeout } from "../../helpers/fetch.js";
 import { richErrorMessage } from "../../helpers/richErrorMessage.js";
 import { useLogger } from "../../logger.js";
 import { VideoError } from "../../errors/VideoError.js";
 
 const logger = useLogger();
 
+const metaSource: Readonly<VideoMetaSource> = {
+	platformName: "soundcloud",
+	alternative: null,
+};
+
 /**
  * Gets information about a SoundCloud track.
  *
  * @param url The track URL to check.
+ * @param signal A signal that would indicate that we should abort the network request.
  *
  * @throws an error if SoundCloud song information could not be found at the
  * provided `url`.
  * @returns a `Promise` that resolves with the track details.
  */
-export async function getSoundCloudTrack(url: URL): Promise<VideoDetails> {
+export async function getSoundCloudTrack(url: URL, signal?: AbortSignal): Promise<VideoDetails> {
 	// TRY:
 	// 1. Use an OpenGraph parser like https://iplocation.io/open-graph-checker
 	// 2. Use og:url in the SoundCloud API client to get duration, use og:title for title
@@ -26,7 +31,7 @@ export async function getSoundCloudTrack(url: URL): Promise<VideoDetails> {
 	// (*.app.goo.gl links come from the app, and redirect to the song page)
 	let parsedUrl: URL;
 	try {
-		const response = await fetchWithTimeout(url.href, undefined, { redirect: "follow" });
+		const response = await fetch(url.href, { redirect: "follow", signal });
 		parsedUrl = new URL(response.url);
 	} catch (error) {
 		logger?.debug(
@@ -41,7 +46,7 @@ export async function getSoundCloudTrack(url: URL): Promise<VideoDetails> {
 	const client = new SoundCloudClient();
 	let song: Song;
 	try {
-		song = await client.getSongInfo(parsedUrl.href);
+		song = await client.getSongInfo(parsedUrl.href, { requestOptions: { signal } });
 	} catch (error) {
 		throw new VideoError(error);
 	}
@@ -49,5 +54,6 @@ export async function getSoundCloudTrack(url: URL): Promise<VideoDetails> {
 		url: song.url,
 		title: song.title,
 		duration: { seconds: Math.floor(song.duration / 1000) },
+		metaSource,
 	};
 }