+++ src/commands/__mocks__/index.ts
@@ -7,7 +7,7 @@ interface MockCommand {
 	aliases?: Command["aliases"];
 	nameLocalizations?: Command["nameLocalizations"];
 	options: Command["options"];
-	execute: Mock<Parameters<Command["execute"]>, ReturnType<Command["execute"]>>;
+	execute: Mock<Command["execute"]>;
 }
 
 export { invokeCommand } from "../../actions/invokeCommand.js";
@@ -27,13 +27,9 @@ function addMock(command: Command): void {
 		aliases: command.aliases,
 		options: command.options,
 		nameLocalizations: command.nameLocalizations,
-		execute: vi
-			.fn<Parameters<Command["execute"]>, ReturnType<Command["execute"]>>()
-			.mockResolvedValue(undefined),
+		execute: vi.fn<Command["execute"]>().mockResolvedValue(undefined),
 	});
 }
 
 // Add all commands to our mock commands list
-for (const cmd of _allCommands.values()) {
-	addMock(cmd);
-}
+_allCommands.forEach(addMock);