+++ CHANGELOG.md
@@ -6,6 +6,22 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Changed
+- Updated internal testing utilities.
+
+## [3.1.0] - 2024-08-22
+### Added
+- Use YouTube's first-party API when an API key is provided.
+- The `/test` command now reports when an alternative source is used for querying platforms. This is especially useful to determine whether Gamgee needed to fall back on an Invidius instance when YTDL failed, or when an API key was not configured.
+
+### Changed
+- More reliable parsing of track duration data from Bandcamp, using [a polyfill](https://github.com/fullcalendar/temporal-polyfill) for the new [Temporal API](https://tc39.es/proposal-temporal/docs/duration.html) instead of RegEx.
+
+## [3.0.0] - 2024-08-21
+### Fixed
+- Version bump because of a breaking change in v2.2.1. (Sorry!!) We now require Node 20. Docker users should be unaffected, since the Dockerfile \*should\* be using the latest Node anyway.
+
+## [2.2.1] - 2024-08-21
+### Changed
 - BREAKING: We now require Node 20. Make sure you upgrade your system Node version if you don't use Docker.
 - Network and args parsing are now handled using native structures instead of `cross-fetch` and `yargs`.
 - Replaced `Jest`, `Mocha`, and `Chai` with `Vitest` for unit testing.
@@ -13,6 +29,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Use trailing commas to made code diffs cleaner.
 - `/limits` now mentions the `/cooldown` command properly.
 - Updated dependencies, and made our Rollup bundle tighter.
+
+### Fixed
+- We now halt dangling network requests when one video platform answers back. This might speed up handling request floods, since ostensibly this means Node will let go of unneeded network resources more quickly for each request.
+- Handle YouTube failures by trying an [Invidius](https://docs.invidious.io/api/#get-apiv1videosid) proxy instead. For now, the only instance we'll try is https://iv.ggtyler.dev, as it seemed the most reliable at the time.
 
 ## [2.2.0] - 2024-02-18
 ### Added
@@ -448,7 +468,10 @@ After updating, be sure to run `npm ci && npm run build:clean && npm run migrate
 ### Added
 - Initial commit
 
-[Unreleased]: https://github.com/AverageHelper/Gamgee/compare/v2.2.0...HEAD
+[Unreleased]: https://github.com/AverageHelper/Gamgee/compare/v3.1.0...HEAD
+[3.1.0]: https://github.com/AverageHelper/Gamgee/compare/v3.0.0...v3.1.0
+[3.0.0]: https://github.com/AverageHelper/Gamgee/compare/v2.2.1...v3.0.0
+[2.2.1]: https://github.com/AverageHelper/Gamgee/compare/v2.2.0...v2.2.1
 [2.2.0]: https://github.com/AverageHelper/Gamgee/compare/v2.1.1...v2.2.0
 [2.1.1]: https://github.com/AverageHelper/Gamgee/compare/v2.1.0...v2.1.1
 [2.1.0]: https://github.com/AverageHelper/Gamgee/compare/v2.0.6...v2.1.0